add initial timer and adc support to f4 using device tree setup

+++ src/stm32f4xx/system_stm32f4xx.c
@@ -344,13 +344,13 @@ __libc_init_array (void)
   size_t count;
   size_t i;
 
-  count = __preinit_array_end - __preinit_array_start;
+  count = (size_t)(__preinit_array_end - __preinit_array_start);
   for (i = 0; i < count; i++)
     __preinit_array_start[i] ();
 
   //_init ();
 
-  count = __init_array_end - __init_array_start;
+  count = (size_t)(__init_array_end - __init_array_start);
   for (i = 0; i < count; i++)
     __init_array_start[i] ();
 }
